:recycle: relocate article entity
