+++ src/users/users.service.ts
@@ -13,7 +13,7 @@ import { HashHelper } from 'src/common/hash.helper';
 import { ArticlesService } from '../articles/articles.service';
 import { UsersToArticlesService } from '../userstoarticles/userstoarticles.service';
 import { SaveArticleDto } from '../articles/articles.dto';
-import { Article } from '../articles/articles.entity';
+import { Article } from '../articles/entities/articles.entity';
 import { UsersToArticles } from '../userstoarticles/userstoarticles.entity';
 import { ShareArticleDto } from './dto/shareArticle.dto';
 import { UserBody } from 'src/authentication/dto/userbody.dto';
