+++ src/userstoarticles/userstoarticles.entity.ts
@@ -1,5 +1,5 @@
 import { Entity, CreateDateColumn, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
-import { Article } from '../articles/articles.entity';
+import { Article } from '../articles/entities/articles.entity';
 import { User } from '../users/entities/user.entity';
 
 @Entity()
